[Fix] #13 - Profile 폴더 구조 수정
